Separate out Google Realtime stuff from CanvasElementService
