+++ src/app/graph/secondary-toolbar/secondary-toolbar.ts
@@ -1,8 +1,8 @@
 import {Component} from '@angular/core';
 import {
-  CanvasElementManagerService,
+  CanvasElementService,
   ArrowConnectionType
-} from 'src/app/canvas/canvas-element-manager/canvas-element-manager.service';
+} from 'src/app/canvas/canvas-element.service';
 
 @Component({
   selector: 'uxg-secondary-toolbar',
@@ -11,16 +11,16 @@ import {
 })
 export class SecondaryToolbarComponent {
   constructor(
-    private canvasElementManagerService: CanvasElementManagerService) {
+    private canvasElementService: CanvasElementService) {
   }
 
   onAddCardButtonPressed(event: MouseEvent) {
-    this.canvasElementManagerService.realtimeTransaction(() => {
-      const card = this.canvasElementManagerService.addCard();
-      const arrow = this.canvasElementManagerService.addArrow();
+    this.canvasElementService.transaction(() => {
+      const card = this.canvasElementService.addCard();
+      const arrow = this.canvasElementService.addArrow();
       if (card !== null && arrow !== null) {
-        this.canvasElementManagerService.connectArrowAndCard(arrow, card, ArrowConnectionType.OUTGOING);
-        this.canvasElementManagerService.repositionArrow(arrow);
+        this.canvasElementService.connectArrowAndCard(arrow, card, ArrowConnectionType.OUTGOING);
+        this.canvasElementService.repositionArrow(arrow);
       }
     });
   }
